+++ CHANGELOG.md
@@ -14,9 +14,11 @@ x.x.x - TODO
 
 4.x.x - 2026xxxx 
   * Changes in config/default.php file:
+    - update $config['cops_twig_templates'] to include 'markdown' templates
     - new $config['cops_browse_books_directory'] to browse books in other folders
     - new $config['cops_browse_books_getfiles'] to get list of files in other folders
   * Browse book files in other folders besides Calibre (WIP)
+  * Start adding markdown templates as experiment (WIP)
 
 4.2.0 - 20260209 Support mapping of filepath for x-accel-redirect
   * Changes in config/default.php file: